#874 remove base pouch cell
