--- docs/source/models/submodels/thermal/pouch_cell/base_pouch_cell.rst
@@ -1,5 +1,0 @@
-Base Thermal Model for "N+1D" Pouch Cell
-========================================
-
-.. autoclass:: pybamm.thermal.pouch_cell.BasePouchCell
-    :members:
