+++ docs/source/models/submodels/thermal/pouch_cell/index.rst
@@ -4,6 +4,5 @@ Pouch Cell
 .. toctree::
   :maxdepth: 1
 
-  base_pouch_cell
   pouch_cell_1D_current_collectors
   pouch_cell_2D_current_collectors
